+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Built-in support for encoding and decoding compressed conf as url param.
+- Add heatmap support to schema.
 
 ### Changed
 - Fixed schema validity as state updates.